Add the health_zone routes
